refactor(dto): rename Data namespace to DTOs and update generators
Rename the App\Data namespace to App\DTOs for clarity and update all
imports across controllers and routes to reference the new DTOs
namespace. Adjust DTO class files (BreadCrumb, PageMeta) to the new
namespace. Rename and repurpose the artisan generator command from
MakeDataCommand to MakeDTOCommand, update its command name and
description (make:data -> make:dto), change stub path to dto.stub, and
set the default generation namespace to App\DTOs. Update option help
text to reference "DTO" instead of "data".

These changes make the intent of the value objects clearer (DTO = Data
Transfer Object), keep naming consistent project-wide, and ensure the
code generator produces files in the correct namespace.
