+++ app/Http/Controllers/Auth/RegisteredUserController.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\Http\Controllers\Auth;
 
-use App\Data\PageMeta;
+use App\DTOs\PageMeta;
 use App\Http\Controllers\Controller;
 use App\Models\User;
 use Illuminate\Auth\Events\Registered;
